raspberrypi: add state script to update boot firmware
Raspberry Pi boards have a set of boot firmware files that are
located on the vfat boot part, the files are:

    bootcode.bin  fixup.dat     fixup_cd.dat  fixup_db.dat
    fixup_x.dat   start.elf     start_cd.elf  start_db.elf
    start_x.elf

Occasionally there will be changes to the Raspberry Pi software stack
that requires that these files are update. Typically there is something
in the Linux kernel that requires something that is in the boot
firmware. This means that to update the Linux kernel you must also
update the boot firmware files.

Above is really sub-optimal design of the Raspberry Pi boards but a
limitation that we must live with.

Note that the DTB files are also part of "boot firmware" and included in
the state script update.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>