raspberrypi: add bcm2835-bootfile-target recipe
This recipe is basically a copy of the bcm2835-bootfiles recipe that is
present in meta-raspberrypi, but installing the boot-files in a
directory on target instead of DEPLOYDIR.

I would prefer this to be an bbappend but the upstream recipe inherits
"nopackages" and there is no way to override that in a bbappend.

To make use of it include it in the target image, one way of doing this
is by setting the following in local.conf

   CORE_IMAGE_EXTRA_INSTALL += "bcm2835-bootfiles-target"

This will include the boot firmware files and the state script in the
Mender artifact.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>