rpi: update U-boot patch to get rid of warning
Even though there does not seem to be any problems applying the
current patch it does produce a warning:

WARNING: u-boot-1_2018.01-r0 do_patch:
Some of the context lines in patches were ignored. This can lead to
incorrectly applied patches.
The context lines in the patches can be updated with devtool:

    devtool modify <recipe>
    devtool finish --force-patch-refresh <recipe> <layer_path>

Then the updated patches and the source tree (in devtool's workspace)
should be reviewed to make sure the patches apply in the correct place
and don't introduce duplicate lines (which can, and does happen
when some of the context is ignored). Further information:
http://lists.openembedded.org/pipermail/openembedded-core/2018-March/148675.html
https://bugzilla.yoctoproject.org/show_bug.cgi?id=10450
Details:
Applying patch 0001-CONFIGS-rpi-enable-mender-requirements.patch

Changelog: Title

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>