fixes to be able to build webOS OSE
That is:

   https://github.com/webosose/build-webos

webOS uses a combiniation of morty and rocko branhces of Yocto.

oe-core is morty, so we have to use morty of meta-mender to avoid
parse errors due to syntax changes in getVar command.

webOS though uses U-boot 2017.09 which is from rocko branch, hence
I have simply copied the patches from meta-mender/rocko to meta-mender/morty.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>