MEN-3366: move data part resize to a script
In bd490c8d59f7 ("MEN-3366: write GPT partition table before resize of data part"),
we added an extra command to ensure that GPT backup headers are written
prior to trying to resize the partition.

This means that on each boot we would execute "fdisk w", the downside
of doing this is that it is causing wear of the storage medium but
also there is an risk involved with manipulating the partition table,
especially on each boot. Even if GPT has backup headers (which is
what caused this problem to beging with) it is probably best to
minimize the impact of this feature and isolate it to only run
first boot after provisioning.

This commit tries to remedy above problems by moving the logic to
an external script that checks if data part has been resized, and
only execute the resize operations if they have not been done before.

Changelog: None

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>