add mender-artifact 3.3.0 recipe
This tool is used to create Mender Artifacts which can
utilized to deploy OTA software.

This recipe relies on fetching a binary version of the
mender-artifact instead of trying to compile from source.
mender-artifact is written in GO, and support for GO is
not great in Yocto 1.8.

We might have compiled from source but would probably need
to compile an older version of the tool, with this approach
we can use the latest.

Signed-off-by: Mirza Krak <mirza.krak@northern.tech>